OrderFlow-API-main/api/requirements.txt
Django==3.2.24
djangorestframework==3.12.4
djoser==2.1.0
djangorestframework-simplejwt==4.7.2
psycopg2-binary==2.9.10

# Исправленная версия social-auth-app-django для совместимости с djoser 2.1.0
social-auth-app-django==4.0.0
gunicorn==20.1.0
channels==4.0.0
daphne==4.0.0
uvicorn==0.23.2
channels_redis==4.1.0
openai==1.1.1
langchain==0.0.267
autopep8==2.0.4
black==23.9.1
Pillow==9.5.0
drf-yasg==1.21.7
django-cors-headers==4.3.1
pytest==7.4.2
pytest-django==4.5.2
python-dotenv==1.0.0
django-extensions==3.2.3
pycodestyle==2.10.0
SQLAlchemy==2.0.21
redis==4.6.0
django-redis==5.2.0

# Совместимые версии httpx и зависимостей
httpx==0.27.2

celery==5.3.1
django-celery-beat==2.5.0
django-celery-results==2.5.1

yoomoney==0.1.0
yoomoney 
requests